update python and java code

+++ java/src/DataStructures/MeanStd.java
@@ -1,6 +1,6 @@
 package DataStructures;
 
 public class MeanStd {
-	public double mean;
-	public double std;
+	public float mean;
+	public float std;
 }
